Extract IT-specific functionality from AzulClient
